# refactor: vectorize *Element functions