@sgibb 2021-12-29
> * `countElements` returns now a `list` (of `integer`?) because multiple input formulas are supported, right? If that's the case I would be OK with this incompatible change.

Correct. A `list` of named `integer`s.

> * regarding the `y` parameter in `addElements`: if `y` is of length 1 (e.g. `y = "H"`) this will be added to each formula in `x`, right? And it would be possible to add/substract a different formula from each `x`?

Correct. If multiple `x` and a single `y` are given, `y` is add/subtract from each `x`. If a single `x` and multiple `y` are given, the `x` is recycled/duplicated and each `y` is add/substract to `x`. Before multiple `y` are cumulative added/subtracted to/from a single `x`. For multiple `x` and multiple `y` (given the same length), each `x` is added/subtracted to/from each corresponding `y`.